requirements.txt
ansible-core
PyYAML
pytest
pytest-xdist
pytest-html
testinfra
typer
